WIP: prøve å forbedre opprydning av mocker.
