+++ src/main/kotlin/no/nav/familie/ba/sak/Application.kt
@@ -1,6 +1,5 @@
 package no.nav.familie.ba.sak
 
-import no.nav.familie.ba.sak.config.ApplicationConfig
 import no.nav.familie.ba.sak.config.DelayedShutdownHook
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
@@ -10,7 +9,7 @@ import org.springframework.context.ConfigurableApplicationContext
 class FamilieBaSakApplication
 
 fun main(args: Array<String>) {
-    val app = SpringApplication(ApplicationConfig::class.java)
+    val app = SpringApplication(FamilieBaSakApplication::class.java)
     app.setRegisterShutdownHook(false)
     val applicationContext: ConfigurableApplicationContext = app.run(*args)
     Runtime.getRuntime().addShutdownHook(DelayedShutdownHook(applicationContext))
