+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/skyggesak/SkyggesakService.kt
@@ -1,14 +1,12 @@
 package no.nav.familie.ba.sak.integrasjoner.skyggesak
 
 import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.IntegrasjonClient
-import no.nav.familie.ba.sak.integrasjoner.pdl.PersonopplysningerService
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import org.springframework.stereotype.Service
 
 @Service
 class SkyggesakService(
-    private val integrasjonClient: IntegrasjonClient,
-    private val personopplysningerService: PersonopplysningerService
+    private val integrasjonClient: IntegrasjonClient
 ) {
 
     fun opprettSkyggesak(aktør: Aktør, fagsakId: Long) {
