+++ src/test/kotlin/DevLauncher.kt
@@ -1,22 +1,16 @@
-import no.nav.familie.ba.sak.config.ApplicationConfig
 import org.springframework.boot.builder.SpringApplicationBuilder
 
-object DevLauncher {
+class DevLauncher
 
-    @JvmStatic
-    fun main(args: Array<String>) {
-        System.setProperty("spring.profiles.active", "dev")
-        val app = SpringApplicationBuilder(ApplicationConfig::class.java)
-            .profiles(
-                "dev",
-                "mock-brev-klient",
-                "mock-økonomi",
-                "mock-infotrygd-feed",
-                "mock-infotrygd-barnetrygd",
-                "mock-pdl",
-                "mock-tilbakekreving-klient",
-                "task-scheduling"
-            )
-        app.run(*args)
-    }
+fun main(args: Array<String>) {
+    SpringApplicationBuilder(DevLauncher::class.java).profiles(
+        "dev",
+        "mock-brev-klient",
+        "mock-økonomi",
+        "mock-infotrygd-feed",
+        "mock-infotrygd-barnetrygd",
+        "mock-pdl",
+        "mock-tilbakekreving-klient",
+        "task-scheduling"
+    ).run(*args)
 }
