+++ src/test/kotlin/DevLauncherPostgres.kt
@@ -1,10 +1,9 @@
-import no.nav.familie.ba.sak.config.ApplicationConfig
 import org.springframework.boot.builder.SpringApplicationBuilder
 
 class DevLauncherPostgres
 
 fun main(args: Array<String>) {
-    SpringApplicationBuilder(ApplicationConfig::class.java).profiles(
+    SpringApplicationBuilder(DevLauncherPostgres::class.java).profiles(
         "postgres",
         "mock-brev-klient",
         "mock-økonomi",
