+++ src/test/kotlin/no/nav/familie/ba/sak/config/LocalDateServiceTestConfig.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.ba.sak.config
 
+import io.mockk.clearMocks
 import io.mockk.every
 import io.mockk.mockk
 import no.nav.familie.ba.sak.common.LocalDateService
@@ -18,13 +19,15 @@ class LocalDateServiceTestConfig {
     fun mockLocalDateService(): LocalDateService {
         val mockLocalDateService = mockk<LocalDateService>()
 
-        clearMocks(mockLocalDateService)
+        clearLocalDateServiceMocks(mockLocalDateService)
 
         return mockLocalDateService
     }
 
     companion object {
-        fun clearMocks(mockLocalDateService: LocalDateService) {
+        fun clearLocalDateServiceMocks(mockLocalDateService: LocalDateService) {
+            clearMocks(mockLocalDateService)
+
             every { mockLocalDateService.now() } returns LocalDate.now()
         }
     }
